finished part 4

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog.js")
 const User = require("../models/user.js")
+const jwt = require("jsonwebtoken")
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
@@ -11,23 +12,42 @@ blogsRouter.post("/", async (request, response) => {
   if (!("title" in request.body || "url" in request.body))
     return response.status(400).end()
 
-  // if (!("likes" in request.body)) request.body.likes = 0
+  const token = await jwt.verify(request.token, process.env.SECRET)
+  if (request.token === undefined || !token.id)
+    return response.status(401).json({ error: "missing or invalid token" })
 
-  const randomUser = await User.findOne({})
+  const user = await User.findById(token.id)
 
   const blog = new Blog({
     title: request.body.title,
     author: request.body.author,
     url: request.body.url,
     likes: request.body.likes === undefined ? 0 : request.body.likes,
-    user: randomUser.id
+    user: user.id
   })
 
   const result = await blog.save()
+  user.blogs =
+    user.blogs === undefined ? [result.id] : user.blogs.concat(result.id)
+
+  await user.save()
   response.status(201).json(result)
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const token = await jwt.verify(request.token, process.env.SECRET)
+
+  if (request.token === undefined || !token.id)
+    return response.status(401).json({ error: "missing or invalid token" })
+
+  const user = await User.findById(token.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog.user.equals(user.id))
+    return response
+      .status(403)
+      .json({ error: "you don't have the right permissions" })
+
   await Blog.findByIdAndDelete(request.params.id)
   return response.status(204).end()
 })
